% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dEL.R
\name{dEL}
\alias{dEL}
\title{Density function of the Extended Laplace Distribution}
\usage{
dEL(y, mu, sigma, delta)
}
\arguments{
\item{y}{Vector of values where the density is to be evaluated}

\item{mu}{Location parameter}

\item{sigma}{Scale parameter (must be > 0)}

\item{delta}{Uniform noise parameter (must be > 0)}
}
\value{
Vector of density values
}
\description{
Density function of the Extended Laplace Distribution
}
\references{
Saah, D. K., & Kozubowski, T. J. (2025). A new class of extended Laplace distributions with applications to modeling contaminated Laplace data. Journal of Computational and Applied Mathematics. \doi{10.1016/j.cam.2025.116588}
}
